requirements.txt
requests
Pillow
numpy
transformers
torch 
faiss-cpu
Flask
python-dotenv
tqdm
h5py 
scikit-learn
sentence-transformers 